Mock Docker for serial engine notebook unit test
